backend/requirements.txt
fastapi
uvicorn
python-docx
docxcompose
PyMuPDF
sentence-transformers
spacy
scikit-learn
pydantic
python-multipart
